+++ seiscat/fetchdata/event_waveforms.py
@@ -10,6 +10,7 @@ Fetch event waveforms from FDSN web services or from a local SDS archive.
     (https://www.gnu.org/licenses/gpl-3.0-standalone.html)
 """
 import pathlib
+from .sds import get_sds_client, fetch_sds_waveforms
 from .mass_downloader import mass_download_waveforms
 from ..database.dbfunctions import read_events_from_db
 from ..utils import ExceptionExit
@@ -24,7 +25,13 @@ def fetch_event_waveforms(config):
     """
     with ExceptionExit():
         events = read_events_from_db(config)
-    event_dir = pathlib.Path(config['event_dir'])
-    event_dir.mkdir(parents=True, exist_ok=True)
-    for event in events:
-        mass_download_waveforms(config, event)
+        event_dir = pathlib.Path(config['event_dir'])
+        event_dir.mkdir(parents=True, exist_ok=True)
+        args = config['args']
+        if args.sds:
+            sds_client = get_sds_client(args.sds)
+            for event in events:
+                fetch_sds_waveforms(config, event, sds_client)
+        else:
+            for event in events:
+                mass_download_waveforms(config, event)